PuppetMovementFixed

+++ Satan Claus/Assets/Scripts/Puppets/Movement/HandMovement.cs
@@ -13,7 +13,7 @@ public class HandMovement : MonoBehaviour
     private void Awake() {
         rb = GetComponent<Rigidbody2D>();
     }
-    public void Move(int value)
+    public void Move(float value)
     {
          rb.velocity = Vector2.right * speed * value;
     }
